+++ test/test_klf.jl
@@ -2,6 +2,7 @@ module Test_klf
 
 using Random
 using LinearAlgebra
+using SparseArrays
 using MatrixPencils
 using Test
 
@@ -9,6 +10,282 @@ using Test
 Random.seed!(2351);
 
 @testset "Matrix Pencils Utilities" begin
+
+
+@testset "pbalance!" begin
+
+n = 10;  k = 20; 
+for k in (10,20)
+
+lambda = exp.(im.*rand(n));
+la = real(lambda); li = imag(lambda); la[end] = 0; li[end] = 0;  
+D = sort(la[1:end-1]./li[1:end-1])
+
+La = Diagonal(la); Li = Diagonal(li); 
+Tl = randn(n,n).^k; Tr = randn(n,n).^k;
+A = Tl*La*Tr; E = Tl*Li*Tr;
+atol1 = 1.e-4; atol2 = 1.e-4;
+for iter = 1:10
+    ev = sort(pzeros(A,E;atol1,atol2)[1],by=real)
+    length(ev) == n-1 && break
+    atol1 *= 10; atol2 *= 10
+end
+ev = sort(pzeros(A,E;atol1,atol2)[1],by=real)
+corig = norm(abs.(ev-D)./sqrt.(1. .+ ev .^2)./sqrt.(1. .+ D .^2))
+isfinite(corig) || (corig = Inf)
+
+# if any(isinf.(ev))
+#    corig = Inf
+# else
+#    corig = norm(abs.(ev-D)./sqrt.(1. .+ ev .^2)./sqrt.(1. .+ D .^2))
+#    isfinite(corig) || (corig = Inf)
+# end
+
+qsorig = qS1(abs.(A)+abs.(E))
+AA = copy(A); EE = copy(E); 
+@time D1, D2 = regbalance!(AA,EE)
+@test AA == D1*A*D2 && EE == D1*E*D2 
+qsfin = qS1(abs.(AA)+abs.(EE))
+@test qsfin < qsorig 
+atol1 = 1.e-5; atol2 = 1.e-9;
+for iter = 1:10
+    evs = sort(pzeros(AA,EE;atol1,atol2)[1],by=real)
+    length(evs) == n-1 && all(isfinite.(evs)) && break
+    (!all(isfinite.(evs)) || length(evs) > n-1) && (atol1 *= 10; atol2 *= 10)
+    (!all(isfinite.(evs)) || length(evs) < n-1) && (atol1 /= 10; atol2 /= 10)
+end
+evs = sort(pzeros(AA,EE;atol1,atol2)[1],by=real)
+#@test evs ≈ D
+# compute the chordal distance between exact and computed eigenvalues
+cofin = norm(abs.(evs-D)./sqrt.(1. .+ evs .^2)./sqrt.(1. .+ D .^2))
+#println("k = $k cofin = $cofin")
+@test cofin < corig 
+end
+
+
+# Example 5.6 of Dopico et al. SIMAX, 2022
+M = [1. 1 0; 1 0 0;0 0 1];
+MM = copy(M);
+dleft, dright = rcsumsbal!(MM,r = ones(3),c = ones(3))
+@test dleft*M*dright ≈ MM
+@test qS1(MM) < qS1(M)
+# println("qs = $(qS1(M)) qsbal = $(qS1(MM))")
+
+# diagonal regularization
+M = [1. 1 0; 1 0 0;0 0 1];
+α = 0.1
+for α in (1.,0.5,0.1)
+W = [α^2*I M; M' α^2*I] 
+dleft1, dright1 = rcsumsbal!(W,r = ones(6),c = ones(6),tol=0.001)
+Dl = Diagonal(dleft1.diag[1:3]); Dr = Diagonal(dright1.diag[4:6])
+@test qS1(Dl*M*Dr) < qS1(M)
+# println("α = $α  qs = $(qS1(Dl*M*Dr))")
+end
+
+# regularization of Dopico et al. SIMAX, 2022
+M = [1. 1 0; 1 0 0;0 0 1];
+qorig = qS1(M)
+α = 0.1
+for α in (1.,0.5,0.1)
+W = [fill((α/3)^2,3,3) M; M' fill((α/3)^2,3,3)] 
+dleft1, dright1 = rcsumsbal!(W,r = ones(6),c = ones(6),maxiter=1000,tol=0.001)
+Dl = Diagonal(dleft1.diag[1:3]); Dr = Diagonal(dright1.diag[4:6])
+@test qS1(Dl*M*Dr) < qorig
+# println("α = $α  qs = $qorig qsbal = $(qS1(Dl*M*Dr))")
+end
+
+
+α = 0.1
+for α in (1.,0.5,0.1)
+M =  [-1. 0 0; 0 0 0;0 0 -1]; N =  [0. 1 0; 1 0 0;0 0 0];
+MM = copy(M); NN = copy(N)
+dleft, dright = pbalance!(MM, NN; maxiter=1000, tol=0.001, regpar=α, pow2 = false)
+@test dleft*M*dright ≈ MM && dleft*N*dright ≈ NN
+@test qS1(abs.(MM)+abs.(NN)) < qS1(abs.(M)+abs.(N)) 
+# println("diagreg = false α = $α  qs = $(qS1(abs.(MM)+abs.(NN)))")
+
+MM = copy(M); NN = copy(N)
+dleft, dright = pbalance!(MM, NN; diagreg = true, maxiter=1000, tol=0.001, regpar=α, pow2 = false)
+@test dleft*M*dright ≈ MM && dleft*N*dright ≈ NN
+@test qS1(abs.(MM)+abs.(NN)) < qS1(abs.(M)+abs.(N)) 
+# println("diagreg = true α = $α  qs = $(qS1(abs.(MM)+abs.(NN)))")
+
+end
+
+# Example 5.9 of Dopica et al. SIMAX, 2022
+M = [1. 1 1; 0 0 1];
+qorig = qS1(M)
+MM = copy(M);
+dleft, dright = rcsumsbal!(MM,r = 3*ones(2),c = 2*ones(3))
+@test dleft*M*dright ≈ MM
+@test qS1(MM) < qorig
+# println("qs = $qorig qsbal = $(qS1(MM))")
+
+M = [1. 1 1; 0 0 1];
+qorig = qS1(M)
+m, n = size(M)
+α = 1.e-10
+for α in (0.5,0.1,0.01,1.e-4,1.e-10)
+W = [α^2*I M; M' α^2*I] 
+WW = copy(W)
+v = [3*ones(2);2*ones(3)]
+dleft1, dright1 = rcsumsbal!(WW, r = v, c = v, maxiter=1000,tol=0.001)
+Dl = Diagonal(dleft1.diag[1:m]); Dr = Diagonal(dright1.diag[m+1:m+n])
+@test qS1(Dl*M*Dr) < qorig 
+# println("α = $α  qs = $qorig qsbal = $(qS1(Dl*M*Dr))")
+end
+
+
+M = [1. 1 1; 0 0 1];
+qorig = qS1(M)
+m, n = size(M)
+α = 1.e-10
+for α in (0.5,0.1,0.01,1.e-4,1.e-10)
+# W = [α*I M; M' α*I] 
+# dleft, dright = rcsumsbal!(W, r = [3*ones(2);2*ones(3)], c = [3*ones(2);2*ones(3)],tol=0.001)
+W = [fill((α/m)^2,m,m) M; M' fill((α/n)^2,n,n)] 
+WW = copy(W)
+v = [3*ones(2);2*ones(3)]
+dleft1, dright1 = rcsumsbal!(WW,r = v, c = v, maxiter=1000,tol=0.001)
+Dl = Diagonal(dleft1.diag[1:m]); Dr = Diagonal(dright1.diag[m+1:m+n])
+@test qS1(Dl*M*Dr) < qorig 
+# println("α = $α  qs = $qorig qsbal = $(qS1(Dl*M*Dr))")
+end
+
+
+α = 0.1
+M =  [0. -1 -1; 0 0 0]; N =  [1. 0 0; 0 0 1];
+r = 3. *ones(2); c = 2. *ones(3); 
+for α in (0.5,0.1,0.01,1.e-4,1.e-10)
+MM = copy(M); NN = copy(N)
+dleft, dright = pbalance!(MM, NN; r, c, maxiter=1000, tol=0.001, regpar=α, pow2 = false)
+@test dleft*M*dright ≈ MM && dleft*N*dright ≈ NN
+@test qS1(abs.(MM)+abs.(NN)) < qS1(abs.(M)+abs.(N)) 
+# println("diagreg = false α = $α  qs = $(qS1(abs.(MM)+abs.(NN)))")
+
+MM = copy(M); NN = copy(N)
+dleft, dright = pbalance!(MM, NN; diagreg = true, r, c, maxiter=1000, tol=0.001, regpar=α, pow2 = false)
+@test dleft*M*dright ≈ MM && dleft*N*dright ≈ NN
+@test qS1(abs.(MM)+abs.(NN)) < qS1(abs.(M)+abs.(N)) 
+# println("diagreg = true α = $α  qs = $(qS1(abs.(MM)+abs.(NN)))")
+
+end
+
+
+
+
+
+# examples from Dopico et al. SIMAX, 43:1213-1237, 2022.    
+n = 50; k = -16; 
+for k in (-16 : 5 : 15)
+
+T = rand(n,n);
+T[1, 2 : n] = 10. ^(k)*T[1, 2 : n]
+T[4:n, 3] = 10. ^(k)*T[4:n, 3]
+D = sort(round.(Int,1. ./ rand(n)))
+A = T*Diagonal(D); E = T;
+ev = sort(eigvals(A,E),by=real)
+# println("norm(evs-ev) = $(norm(ev - D))")
+# compute the chordal distance between exact and computed eigenvalues
+corig = norm(abs.(ev-D)./sqrt.(1. .+ ev .^2)./sqrt.(1. .+ D .^2))
+isfinite(corig) || (corig = Inf)
+# println("k = $k corig = $corig")
+
+M = abs.(A)+abs.(E); M[2,:] .= 0; M[4,:] .= 0; M[:,4] .= 0;
+MM = copy(M);
+dleft, dright = rcsumsbal!(MM)
+@test dleft*M*dright ≈ MM
+@test qS1(MM) < qS1(M)
+
+
+qsorig = qS1(abs.(A)+abs.(E))
+AA = copy(A); EE = copy(E); 
+@time D1, D2 = pbalance!(AA,EE)
+@test AA == D1*A*D2 && EE == D1*E*D2 
+qsfin = qS1(abs.(AA)+abs.(EE))
+@test qsfin < qsorig 
+#println("qsorig/qsfin = $(qsorig/qsfin)")
+@test 2. .^round.(Int,log2.(D1.diag)) == D1.diag && 2. .^round.(Int,log2.(D2.diag)) == D2.diag
+evs = sort(eigvals(AA,EE),by=real)
+@test evs ≈ D
+# compute the chordal distance between exact and computed eigenvalues
+cofin = norm(abs.(evs-D)./sqrt.(1. .+ evs .^2)./sqrt.(1. .+ D .^2))
+# println("k = $k pow2 = true cofin = $cofin")
+@test ev ≈ D ? cofin/10 < corig :  cofin < corig 
+
+
+AA = copy(A); EE = copy(E); 
+@time D1, D2 = pbalance!(AA,EE; pow2 = false)
+@test AA ≈ D1*A*D2 && EE ≈ D1*E*D2 
+qsfin = max(qS1(AA),qS1(EE))
+@test qsfin < qsorig 
+#println("qsorig/qsfin = $(qsorig/qsfin)")
+evs = sort(eigvals(AA,EE),by=real)
+@test evs ≈ D
+cofin = norm(abs.(evs-D)./sqrt.(1. .+ evs .^2)./sqrt.(1. .+ D .^2))
+# println("k = $k pow2 = false cofin = $cofin")
+end
+
+#  Example 1, Ward 1981, pp.148f 
+
+A = [-2.0e+1 -1.0e+4 -2.0e+0 -1.0e+6 -1.0e+1 -2.0e+5
+6.0e-3 4.0e+0 6.0e-4 2.0e+2 3.0e-3 3.0e+1
+-2.0e-1 -3.0e+2 -4.0e-2 -1.0e+4 0.0e+0 3.0e+3
+6.0e-5 4.0e-2 9.0e-6 9.0e+0 3.0e-5 5.0e-1
+6.0e-2 5.0e+1 8.0e-3 -4.0e+3 8.0e-2 0.0e+0
+0.0e+0 1.0e+3 7.0e-1 -2.0e+5 1.3e+1 -6.0e+4 ];
+
+E = [-2.0e+1 -1.0e+4 2.0e+0 -2.0e+6 1.0e+1 -1.0e+5
+5.0e-3 3.0e+0 -2.0e-4 4.0e+2 -1.0e-3 3.0e+1
+0.0e+0 -1.0e+2 -8.0e-2 2.0e+4 -4.0e-1 0.0e+0
+5.0e-5 3.0e-2 2.0e-6 4.0e+0 2.0e-5 1.0e-1
+4.0e-2 3.0e+1 -1.0e-3 3.0e+3 -1.0e-2 6.0e+2
+-1.0e+0 0.0e+0 4.0e-1 -1.0e+5 4.0e+0 2.0e+4 ]; 
+
+D = collect(1.:1.:6.)  # exact eigenvalues
+
+qsorig = qS1(abs.(A)+abs.(E))
+AA = copy(A); EE = copy(E); 
+@time D1, D2 = pbalance!(AA,EE; pow2 = false)
+@test AA ≈ D1*A*D2 && EE ≈ D1*E*D2 
+qsfin = qS1(abs.(AA)+abs.(EE))
+ev = sort(eigvals(A,E),by=real)
+corig = norm(abs.(ev-D)./sqrt.(1. .+ ev .^2)./sqrt.(1. .+ D .^2))
+# println("corig = $corig")
+evs = sort(eigvals(AA,EE),by=real)
+cofin = norm(abs.(evs-D)./sqrt.(1. .+ evs .^2)./sqrt.(1. .+ D .^2))
+# println("cofin = $cofin")
+@test 100000*cofin < corig && 100000*qsfin < qsorig 
+
+
+# Example 3 (graded matrix), see Ward, 1981, pp.148f
+
+A = [    1.0000e+00   1.0000e+01            0            0            0            0            0
+1.0000e+01   1.0000e+02   1.0000e+03            0            0            0            0
+         0   1.0000e+03   1.0000e+04   1.0000e+05            0            0            0
+         0            0   1.0000e+05   1.0000e+06   1.0000e+07            0            0
+         0            0            0   1.0000e+07   1.0000e+08   1.0000e+09            0
+         0            0            0            0   1.0000e+09   1.0000e+10   1.0000e+11
+         0            0            0            0            0   1.0000e+11   1.0000e+12];
+E = 1. *Matrix(I(7)); 
+qsorig = qS1(abs.(A)+abs.(E))
+AA = copy(A); EE = copy(E); 
+@time D1, D2 = pbalance!(AA,EE; pow2 = false)
+@test AA ≈ D1*A*D2 && EE ≈ D1*E*D2 
+qsfin = qS1(abs.(AA)+abs.(EE))
+@test 100000*qsfin < qsorig 
+
+As = sparse(A); Es = sparse(E)
+qsorig = qS1(abs.(As)+abs.(Es))
+AA = copy(As); EE = copy(Es); 
+@time D1, D2 = pbalance!(AA,EE; pow2 = false)
+@test AA ≈ D1*As*D2 && EE ≈ D1*Es*D2 
+qsfin = qS1(abs.(AA)+abs.(EE))
+@test 100000*qsfin < qsorig 
+
+
+end
+
 
 println("klf_rlsplit")      
 @testset "klf_rlsplit" begin